added policy index page

+++ database/migrations/2019_01_04_090435_create_checkin_totals_table.php
@@ -16,9 +16,12 @@ class CreateCheckinTotalsTable extends Migration
         Schema::create('checkin_totals', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('child_id');
-            $table->float('am_total_hours')->default(0);
-            $table->float('total_hours')->default(0);
-            $table->float('total_amount')->default(0);
+            // $table->float('am_total_hours')->default(0);
+            // $table->float('total_hours')->default(0);
+            // $table->float('total_amount')->default(0);
+            $table->integer('am_total_hours')->default(0);
+            $table->integer('total_hours')->default(0);
+            $table->integer('total_amount')->default(0);
             $table->timestamps();
         });
     }
